pip_requirements.txt
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
einops==0.8.1
filelock==3.13.1
flash-attn==2.3.6
fsspec==2024.6.1
gpg==1.16.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.4
numpy==1.26.4
oauthlib==3.2.0
packaging==24.2
paho-mqtt==2.1.0
pillow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sympy==1.13.1
systemd-python==234
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
triton==2.1.0
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
